backend/requirements.txt
django
djangorestframework
web3
